% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitimeline.R
\name{buildtimeline}
\alias{buildtimeline}
\title{Retrieve and arrange biographical data of several people.}
\usage{
buildtimeline(names, groups = NA)
}
\arguments{
\item{names}{a vector or list of names for which biographical dates should be retrieved}

\item{groups}{passed on to \code{\link{grouptimeperiods}}}
}
\value{
a data frame with one row per name and four columns (name, group, birth, death)
}
\description{
Retrieve and arrange biographical data of several people.
}
\examples{
buildtimeline(c("Hieronymus Bosch", "Michelangelo", "Pieter Bruegel the Elder"))
}
