% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitimeline.R
\name{getwikibio}
\alias{getwikibio}
\title{Retrieve birth and death year of a person from Wikipedia.}
\usage{
getwikibio(name, wiki = "en.wikipedia.org")
}
\arguments{
\item{name}{the name of the Wikipedia article}

\item{wiki}{domain of the wiki to retrieve the article from}
}
\description{
Retrieves the Wikipedia article with the given name and attempts to extract
birth and death data from its \emph{infobox} template.
}
\examples{
getwikibio("Hieronymus Bosch")
}
