% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitimeline.R
\name{grouptimeperiods}
\alias{grouptimeperiods}
\title{Arrange the given time periods into groups of non-overlapping time spans.}
\usage{
grouptimeperiods(starts, ends, groups = NULL)
}
\arguments{
\item{starts}{vector of all start dates of the time periods}

\item{ends}{end times of the time periods (same length as the \code{starts} argument)}

\item{groups}{vector of pre-assigned groups for some of the time periods (optional)}
}
\value{
a character vector of groups, with one item for every time period
}
\description{
Arrange the given time periods into groups of non-overlapping time spans.
}
